move player class to own file
