+++ SignallingWebServer/cirrus.js
@@ -12,6 +12,7 @@ const bodyParser = require('body-parser');
 const logging = require('./modules/logging.js');
 const WebSocket = require('ws');
 const Matchmaker = require('./modules/matchmaker.js');
+const Player = require('./modules/player.js');
 logging.RegisterConsoleLogger();
 
 
@@ -277,56 +278,6 @@ console.logColor(logging.Cyan, `Running Cirrus - The Pixel Streaming reference i
 let nextPlayerId = 1;
 
 const PlayerType = { Regular: 0, SFU: 1 };
-
-class Player {
-	constructor(id, ws, type, browserSendOffer) {
-		this.id = id;
-		this.ws = ws;
-		this.type = type;
-		this.browserSendOffer = browserSendOffer;
-	}
-
-	subscribe(streamerId) {
-		if (!streamers.has(streamerId)) {
-			console.error(`subscribe: Player ${this.id} tried to subscribe to a non-existent streamer ${streamerId}`);
-			return;
-		}
-		this.streamerId = streamerId;
-		const msg = { type: 'playerConnected', playerId: this.id, dataChannel: true, sfu: this.type == PlayerType.SFU, sendOffer: !this.browserSendOffer };
-		logOutgoing(this.streamerId, msg);
-		this.sendFrom(msg);
-	}
-
-	unsubscribe() {
-		if (this.streamerId && streamers.has(this.streamerId)) {
-			const msg = { type: 'playerDisconnected', playerId: this.id };
-			logOutgoing(this.streamerId, msg);
-			this.sendFrom(msg);
-		}
-		this.streamerId = null;
-	}
-
-	sendFrom(message) {
-		if (!this.streamerId) {
-			return;
-		}
-
-		message.playerId = this.id;
-		const msgString = JSON.stringify(message);
-
-		let streamer = streamers.get(this.streamerId);
-		if (!streamer) {
-			console.error(`sendFrom: Player ${this.id} subscribed to non-existent streamer: ${this.streamerId}`);
-		} else {
-			streamer.ws.send(msgString);
-		}
-	}
-
-	sendTo(message) {
-		const msgString = JSON.stringify(message);
-		this.ws.send(msgString);
-	}
-};
 
 let streamers = new Map();		// streamerId <-> streamer socket
 let players = new Map(); 		// playerId <-> player, where player is either a web-browser or a native webrtc player
@@ -622,7 +573,7 @@ sfuServer.on('connection', function (ws, req) {
 			console.error(`ERROR: ws.on error: ${err.message}`);
 		}
 	});
-
+	/** @type {Player} */
 	let sfuPlayer = new Player(SFUPlayerId, ws, PlayerType.SFU, false);
 	players.set(SFUPlayerId, sfuPlayer);
 	console.logColor(logging.Green, `SFU (${req.connection.remoteAddress}) connected `);
@@ -717,6 +668,7 @@ playerServer.on('connection', function (ws, req) {
 	++playerCount;
 	let playerId = sanitizePlayerId(nextPlayerId++);
 	console.logColor(logging.Green, `player ${playerId} (${req.connection.remoteAddress}) connected`);
+	/** @type {Player} */
 	let player = new Player(playerId, ws, PlayerType.Regular, browserSendOffer);
 	players.set(playerId, player);
 	matchmaker?.setPlayers(players);
